add upload and downlaod

+++ lib/segment/id.go
@@ -27,24 +27,24 @@ func (bm *BaseSegmentID) GetFsID() []byte {
 	return bm.buf[:FSID_LEN]
 }
 
-func (bm *BaseSegmentID) GetBucketID() int64 {
-	return int64(binary.BigEndian.Uint64(bm.buf[FSID_LEN : FSID_LEN+8]))
+func (bm *BaseSegmentID) GetBucketID() uint64 {
+	return binary.BigEndian.Uint64(bm.buf[FSID_LEN : FSID_LEN+8])
 }
 
-func (bm *BaseSegmentID) GetStripeID() int64 {
-	return int64(binary.BigEndian.Uint64(bm.buf[FSID_LEN+8 : FSID_LEN+16]))
+func (bm *BaseSegmentID) GetStripeID() uint64 {
+	return binary.BigEndian.Uint64(bm.buf[FSID_LEN+8 : FSID_LEN+16])
 }
 
 func (bm *BaseSegmentID) GetChunkID() uint32 {
 	return binary.BigEndian.Uint32(bm.buf[FSID_LEN+16 : SEGMENTID_LEN])
 }
 
-func (bm *BaseSegmentID) SetBucketID(bid int64) {
-	binary.BigEndian.PutUint64(bm.buf[FSID_LEN:FSID_LEN+8], uint64(bid))
+func (bm *BaseSegmentID) SetBucketID(bid uint64) {
+	binary.BigEndian.PutUint64(bm.buf[FSID_LEN:FSID_LEN+8], bid)
 }
 
-func (bm *BaseSegmentID) SetStripeID(sid int64) {
-	binary.BigEndian.PutUint64(bm.buf[FSID_LEN+8:FSID_LEN+16], uint64(sid))
+func (bm *BaseSegmentID) SetStripeID(sid uint64) {
+	binary.BigEndian.PutUint64(bm.buf[FSID_LEN+8:FSID_LEN+16], sid)
 }
 
 func (bm *BaseSegmentID) SetChunkID(cid uint32) {
@@ -70,7 +70,7 @@ func (bm *BaseSegmentID) IndexString() string {
 	return base58.Encode(bm.IndexBytes())
 }
 
-func NewSegmentID(fid []byte, bid, sid int64, cid uint32) (SegmentID, error) {
+func NewSegmentID(fid []byte, bid, sid uint64, cid uint32) (SegmentID, error) {
 	if len(fid) != FSID_LEN {
 		return nil, ErrWrongKeyLength
 	}
@@ -78,8 +78,8 @@ func NewSegmentID(fid []byte, bid, sid int64, cid uint32) (SegmentID, error) {
 	segID := make([]byte, SEGMENTID_LEN)
 	copy(segID[:FSID_LEN], fid)
 
-	binary.BigEndian.PutUint64(segID[FSID_LEN:FSID_LEN+8], uint64(bid))
-	binary.BigEndian.PutUint64(segID[FSID_LEN+8:FSID_LEN+16], uint64(sid))
+	binary.BigEndian.PutUint64(segID[FSID_LEN:FSID_LEN+8], bid)
+	binary.BigEndian.PutUint64(segID[FSID_LEN+8:FSID_LEN+16], sid)
 	binary.BigEndian.PutUint32(segID[FSID_LEN+16:SEGMENTID_LEN], cid)
 
 	return &BaseSegmentID{buf: segID}, nil
